extract credentials.tsx helpers + components into credentials/ modules
